Changed IndexFromPosition to solve problem in SourcePixelled mapping + changed SourcePixelled constructor

+++ include/Tree.h
@@ -250,7 +250,7 @@ namespace Utilities{
 
 	long IndexFromPosition(double *x,long Npixels,double range,double *center);
 	void PositionFromIndex(unsigned long i,double *x,long Npixels,double range,double *center);
-	int IndexFromPosition(double x,long Npixels,double range,double center);
+	long IndexFromPosition(double x,long Npixels,double range,double center);
 	//inline float isLeft( Point *p0, Point *p1, double *x );
 
 	// isLeft(): tests if a point is Left|On|Right of an infinite line.
